Added documentation.

+++ src/MrKWatkins.OakIO.ZXSpectrum.Tests/Snapshot/Sna/SnaToZ80ConverterTests.cs
@@ -36,27 +36,27 @@ public sealed class SnaToZ80ConverterTests
         z80.Should().BeOfType<Z80V2File>();
 
         z80.Header.BorderColour.Should().Equal(ZXColour.Red);
-        z80.Header.InterruptMode.Should().Equal((byte)2);
+        z80.Header.InterruptMode.Should().Equal(2);
         z80.Header.IFF2.Should().BeTrue();
         z80.Header.InterruptFlipFlop.Should().BeTrue();
 
-        z80.Registers.AF.Should().Equal((ushort)0x1234);
-        z80.Registers.BC.Should().Equal((ushort)0x5678);
-        z80.Registers.DE.Should().Equal((ushort)0x9ABC);
-        z80.Registers.HL.Should().Equal((ushort)0xDEF0);
-        z80.Registers.IX.Should().Equal((ushort)0x1111);
-        z80.Registers.IY.Should().Equal((ushort)0x2222);
-        z80.Registers.PC.Should().Equal((ushort)0x8000);
-        z80.Registers.SP.Should().Equal((ushort)0xFF00);
-        z80.Registers.IR.Should().Equal((ushort)0x3F00);
-        z80.Registers.Shadow.AF.Should().Equal((ushort)0xAAAA);
-        z80.Registers.Shadow.BC.Should().Equal((ushort)0xBBBB);
-        z80.Registers.Shadow.DE.Should().Equal((ushort)0xCCCC);
-        z80.Registers.Shadow.HL.Should().Equal((ushort)0xDDDD);
+        z80.Registers.AF.Should().Equal(0x1234);
+        z80.Registers.BC.Should().Equal(0x5678);
+        z80.Registers.DE.Should().Equal(0x9ABC);
+        z80.Registers.HL.Should().Equal(0xDEF0);
+        z80.Registers.IX.Should().Equal(0x1111);
+        z80.Registers.IY.Should().Equal(0x2222);
+        z80.Registers.PC.Should().Equal(0x8000);
+        z80.Registers.SP.Should().Equal(0xFF00);
+        z80.Registers.IR.Should().Equal(0x3F00);
+        z80.Registers.Shadow.AF.Should().Equal(0xAAAA);
+        z80.Registers.Shadow.BC.Should().Equal(0xBBBB);
+        z80.Registers.Shadow.DE.Should().Equal(0xCCCC);
+        z80.Registers.Shadow.HL.Should().Equal(0xDDDD);
 
         var z80Memory = new byte[65536];
         z80.TryLoadInto(z80Memory).Should().BeTrue();
-        z80Memory[0x8000].Should().Equal((byte)0xF3);
-        z80Memory[0x8001].Should().Equal((byte)0xAF);
+        z80Memory[0x8000].Should().Equal(0xF3);
+        z80Memory[0x8001].Should().Equal(0xAF);
     }
 }
